+++ src/main/java/com/udacity/jwdnd/course1/cloudstorage/controllers/FileController.java
@@ -3,7 +3,9 @@ package com.udacity.jwdnd.course1.cloudstorage.controllers;
 import com.udacity.jwdnd.course1.cloudstorage.models.File;
 import com.udacity.jwdnd.course1.cloudstorage.services.FileService;
 import org.apache.coyote.Response;
+import org.springframework.core.io.ByteArrayResource;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -13,10 +15,7 @@ import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBo
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletResponse;
-import java.io.ByteArrayInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
+import java.io.*;
 
 @Controller
 @RequestMapping
@@ -29,6 +28,7 @@ public class FileController {
     }
 
     @GetMapping("/file/{id}")
+    @ExceptionHandler(FileNotFoundException.class)
     public StreamingResponseBody getFile(HttpServletResponse response, @PathVariable String id) throws IOException{
 
         File file = fileService.getFileById(id);
@@ -46,28 +46,42 @@ public class FileController {
 
     @PostMapping("/file-upload")
     public String fileUpload(@RequestParam("fileUpload") MultipartFile fileUpload, File file, Model model) throws IOException {
-        file.setFilename(fileUpload.getOriginalFilename());
-        file.setContenttype(fileUpload.getContentType());
-        file.setFilesize(String.valueOf(fileUpload.getSize()));
-        file.setFiledata(fileUpload.getBytes());
-
-
-        //if not in system, save or update
-        if(fileService.createFile(file) > -1){
+        //check against uploading files with name already in DB.
+        String filename = fileUpload.getOriginalFilename();
+        String userid = String.valueOf(file.getUserid());
+        if(fileService.getFileByNameByUser(filename, userid) != null){
+            model.addAttribute("error", "File with same name already in system.  Please rename the file");
+            return "result";
+        }
+        try {
+            file.setFilename(fileUpload.getOriginalFilename());
+            file.setContenttype(fileUpload.getContentType());
+            file.setFilesize(String.valueOf(fileUpload.getSize()));
+            file.setFiledata(fileUpload.getBytes());
+            fileService.createFile(file);
             model.addAttribute("success", "Successfully created file");
-        }
-        else {
+        }catch(IOException e){
             model.addAttribute("error", "Error creating the file.");
+        }finally {
+            return "result";
         }
-        return "result";
 
     }
 
     @GetMapping("/deleteFile/{id}")
-    public String deleteFile(@PathVariable String id){
-        //get request param, look up in service, then delete if exists
-        fileService.deleteFile(id);
-
-        return "result";
+    public String deleteFile(@PathVariable String id, Model model){
+        if(id == ""){
+            model.addAttribute("error", "Error with trying to delete the file.  Please try again later.");
+            return "result";
+        }
+        try {
+            //get request param, look up in service, then delete if exists
+            fileService.deleteFile(id);
+            model.addAttribute("success", "Successfully deleted file.");
+        }catch(Exception e){
+            model.addAttribute("error", "There was an issue trying to delete. Please try later");
+        }finally {
+            return "result";
+        }
     }
 }
